% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/longTAPIO_MLD.R
\name{cutree_clusMLD}
\alias{cutree_clusMLD}
\title{cutting tree function for clusterMLD
(Analogous to hclust::cutree)}
\usage{
cutree_clusMLD(output, k = NULL, h = NULL)
}
\arguments{
\item{output}{clusterMLD object returned by a call to `clusterMLD::LongDataCluster()`}

\item{k}{an integer scalar with the desired number of groups}

\item{h}{not used at the moment}
}
\value{
`cutree_clusMLD` returns a vector with group memberships
}
\description{
cutting tree function for clusterMLD
(Analogous to hclust::cutree)
}
