% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importance.R
\name{importance}
\alias{importance}
\title{Cluster-specific features importance}
\usage{
importance(res)
}
\arguments{
\item{res}{The object returned by TAPIO.}
}
\value{
Cluster-specific feature importance
}
\description{
Cluster-specific features importance
}
\examples{
NaN

}
