% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/longTAPIO_trajectories.R
\name{longTAPIO_trajectories}
\alias{longTAPIO_trajectories}
\title{Longitudinal hierarchical clustering with an ensemble of PCA trees
(Trajectories strategy)}
\usage{
longTAPIO_trajectories(
  DATA,
  user_id = NULL,
  k = NaN,
  n_features = NaN,
  n_trees = 500,
  do.pca = TRUE,
  do.MFA = FALSE,
  do.leveling = TRUE,
  levels = 10,
  max.k = 10,
  verbose = FALSE
)
}
\arguments{
\item{DATA}{The input data (rows: samples, columns: features)}

\item{user_id}{Identifiers to group the samples by repeated measures}

\item{k}{Number of clusters.}

\item{n_features}{Number of features to sample (default: sqrt(ncolums))}

\item{n_trees}{Number of trees to grow}

\item{do.pca}{Dimension reduction using PCA (default=TRUE)}

\item{do.MFA}{Multi-View clustering (default=FALSE)}

\item{do.leveling}{Leveling (default=TRUE)}

\item{levels}{Number of levels to cut the dendrogram}

\item{max.k}{Number of maximum clusters when k=NaN}

\item{verbose}{Print details (default=FALSE)}
}
\value{
The cluster solution
}
\description{
Longitudinal hierarchical clustering with an ensemble of PCA trees
(Trajectories strategy)
}
\examples{
NaN

}
